Move ClassToBeInstrumented to the test resources
